Split datasets.repositories module
